requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
idna==2.9
idna-ssl==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pkg-resources==0.0.0
pylint==2.4.4
requests==2.23.0
rope==0.16.0
six==1.14.0
soupsieve==2.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
